+++ src/Models/MultiPopLasso.cpp
@@ -214,13 +214,13 @@ MatrixXf MultiPopLasso::proximal_operator(MatrixXf in, float lr) {
     return (in.array()*sign.array()).matrix();
 }
 
-//MatrixXf MultiPopLasso::deriveMatrixA(double lr, long loops, double tol) {
+//MatrixXf MultiPopLasso::deriveMatrixA(float lr, long loops, float tol) {
 //    long r = beta.rows();
 //    long c = C.rows();
 //    MatrixXf A = MatrixXf::Zero(r, c);
 //    MatrixXf bct = beta*C.transpose();
-//    double prev_residue = numeric_limits<double>::max();
-//    double curr_residue;
+//    float prev_residue = numeric_limits<float>::max();
+//    float curr_residue;
 //    for (long i=0;i<loops;i++){
 //        A = A - lr*(bct - A);
 //        A = project(A);
@@ -321,7 +321,7 @@ MultiPopLasso::MultiPopLasso() {
     lambda = default_lambda;
     mu = default_mu;
     gamma = default_gamma;
-    betaAll = MatrixXd::Ones(1,1);
+    betaAll = MatrixXf::Ones(1,1);
     logisticFlag = false;
 }
 
@@ -342,7 +342,7 @@ MultiPopLasso::MultiPopLasso(const unordered_map<string, string> &options) {
     } catch (std::out_of_range& oor) {
         gamma = default_gamma;
     }
-    betaAll = MatrixXd::Ones(1,1);
+    betaAll = MatrixXf::Ones(1,1);
     logisticFlag = false;
 }
